fix(playground room links): remove extra prop
